+++ docs/CHANGELOG.md
@@ -1,10 +1,10 @@
 Changelog
 =========
 
-1.1.2 (Unreleased)
+1.2.0 (Unreleased)
 -------------------------
+- Enh #57: New group permission "Can manage custom pages"
 - Fix #57: Deny public/guest access to custom page with type "User Account Menu (Settings)"
-
 
 1.1.1 (November 12, 2020)
 -------------------------